pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp-cli"
version = "0.6.7"
description = "A cli for the Model Context Provider"
requires-python = ">=3.11"
readme = "README.md"
authors = [
  { name = "Chris Hay", email = "chrishayuk@somejunkmailbox.com" }
]
keywords = ["llm", "openai", "claude", "mcp", "cli"]
license = {text = "MIT"}
dependencies = [
  "asyncio>=3.4.3",
  "chuk-llm[all]>=0.9.8",
  "chuk-mcp[full,pydantic,ui-manager]>=0.5",
  "chuk-term",
  "chuk-tool-processor>=0.6.12",
  "jsonschema>=4.25.1",
  "prompt-toolkit>=3.0.50",
  "python-dotenv>=1.0.1",
  "rich>=13.9.4",
  "typer>=0.15.2",
]


[project.scripts]
mcp-cli = "mcp_cli.main:app"
mcp-llm = "mcp_cli.llm.__main__:main"

[project.optional-dependencies]
wasm = []
dev = [
  "numpy>=2.2.3",
  "pytest-asyncio>=0.25.3",
  "asyncio>=3.4.3"
]

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]
include = ["mcp_cli*"]

[tool.uv.sources]
chuk-term = { git = "https://github.com/chrishayuk/chuk-term.git" }
[dependency-groups]
dev = [
  "colorama>=0.4.6",
  "pydantic>=2.10.2",
  "pytest-asyncio>=0.25.3",
  "pytest-cov>=6.2.1",
  "ruff>=0.12.10",
]
